+++ ikea-zigbee-drivers/E1743.groovy
@@ -10,7 +10,7 @@ import groovy.time.TimeCategory
 import groovy.transform.Field
 
 @Field static final String DRIVER_NAME = "IKEA Tradfri On/Off Switch (E1743)"
-@Field static final String DRIVER_VERSION = "3.5.0"
+@Field static final String DRIVER_VERSION = "3.5.1"
 
 // Fields for capability.HealthCheck
 @Field static final Map<String, String> HEALTH_CHECK = [
@@ -69,8 +69,8 @@ metadata {
 
 // Called when the device is first added
 def installed() {
-    Log.info "Installing device ..."
-    Log.warn "[IMPORTANT] For battery-powered devices, make sure that you keep your device as close as you can to your Hubitat hub (or any other Zigbee router device) for at least 20 seconds. Otherwise it will successfully pair but it won't work properly!"
+    Log.warn "Installing device ..."
+    Log.warn "[IMPORTANT] For battery-powered devices, make sure that you keep your device as close as you can (less than 2inch / 5cm) to your Hubitat hub for at least 30 seconds. Otherwise the device will successfully pair but it won't work properly!"
 }
 
 // Called when the "Save Preferences" button is clicked
@@ -117,7 +117,7 @@ def healthCheck() {
 // capability.Configuration
 // Note: This method is also called when the device is initially installed
 def configure(auto = false) {
-    Log.info "Configuring device${auto ? " (auto)" : ""} ..."
+    Log.warn "Configuring device${auto ? " (auto)" : ""} ..."
     if (!auto && device.currentValue("powerSource", true) == "battery") {
         Log.warn '[IMPORTANT] Click the "Configure" button immediately after pushing any button on the device in order to first wake it up!'
     }
@@ -163,16 +163,17 @@ def configure(auto = false) {
     cmds += zigbee.readAttribute(0x0000, [0x0001, 0x0003, 0x0004, 0x0005, 0x000A, 0x4000]) // ApplicationVersion, HWVersion, ManufacturerName, ModelIdentifier, ProductCode, SWBuildID
     Utils.sendZigbeeCommands cmds
 
-    Log.info "Configuration done! Refreshing device current state in 10 seconds ..."
+    Log.info "Configuration done; refreshing device current state in 10 seconds ..."
     runIn(10, "tryToRefresh")
 }
 private autoConfigure() {
+    Log.warn "Detected that this device is not properly configured for this driver version (lastCx != ${DRIVER_VERSION})"
     configure(true)
 }
 
 // Implementation for capability.HealthCheck
 def ping() {
-    Log.info "ping ..."
+    Log.warn "ping ..."
     Utils.sendZigbeeCommands(zigbee.readAttribute(0x0000, 0x0000))
     Log.debug "Ping command sent to the device; we'll wait 5 seconds for a reply ..."
     runIn 5, "pingExecute"
@@ -213,7 +214,7 @@ def push(buttonNumber) {
 // Implementation for capability.Refresh
 def refresh(buttonPress = true) {
     if (buttonPress) {
-        Log.info "Refreshing current state ..."
+        Log.warn "Refreshing device current state ..."
         if (device.currentValue("powerSource", true) == "battery") {
             Log.warn '[IMPORTANT] Click the "Refresh" button immediately after pushing any button on the device in order to first wake it up!'
         }
@@ -320,7 +321,7 @@ def parse(String description) {
         
         // Events for capability.HealthCheck
         case { contains it, [clusterInt:0x0000, attrInt:0x0000] }:
-            return Log.info("... pong")
+            return Log.warn("... pong")
         
         // Read Attributes Reponse: PowerSource
         case { contains it, [clusterInt:0x0000, commandInt:0x01, attrInt:0x0007] }:
@@ -350,7 +351,7 @@ def parse(String description) {
 
         // Device_annce: Welcome back! let's sync state.
         case { contains it, [endpointInt:0x00, clusterInt:0x0013, commandInt:0x00] }:
-            Log.info "Rejoined the Zigbee mesh! Refreshing device current state in 3 seconds ..."
+            Log.warn "Rejoined the Zigbee mesh; refreshing device state in 3 seconds ..."
             return runIn(3, "tryToRefresh")
 
         // Read Attributes Response (Basic cluster)
@@ -362,7 +363,7 @@ def parse(String description) {
 
         // Mgmt_Leave_rsp
         case { contains it, [endpointInt:0x00, clusterInt:0x8034, commandInt:0x00] }:
-            return Log.info("Device is leaving the Zigbee mesh. See you later, Aligator!")
+            return Log.warn("Device is leaving the Zigbee mesh. See you later, Aligator!")
 
         // Ignore the following Zigbee messages
         case { contains it, [commandInt:0x0A] }:                                       // ZCL: Attribute report we don't care about (configured by other driver)
@@ -378,7 +379,7 @@ def parse(String description) {
         // Unexpected Zigbee message
         // ---------------------------------------------------------------------------------------------------------------
         default:
-            Log.warn "Sent unexpected Zigbee message: description=${description}, msg=${msg}"
+            Log.error "Sent unexpected Zigbee message: description=${description}, msg=${msg}"
     }
 }