+++ ikea-zigbee-drivers/E2006.groovy
@@ -10,7 +10,7 @@ import groovy.time.TimeCategory
 import groovy.transform.Field
 
 @Field static final String DRIVER_NAME = "IKEA Starkvind Air Purifier (E2006)"
-@Field static final String DRIVER_VERSION = "3.5.0"
+@Field static final String DRIVER_VERSION = "3.5.1"
 
 // Fields for capability.PushableButton
 @Field static final List<String> SUPPORTED_FAN_SPEEDS = [
@@ -26,9 +26,10 @@ import groovy.transform.Field
 metadata {
     definition(name:DRIVER_NAME, namespace:"dandanache", author:"Dan Danache", importUrl:"https://raw.githubusercontent.com/dan-danache/hubitat/master/ikea-zigbee-drivers/E2006.groovy") {
         capability "Configuration"
-        capability "Switch"
+        capability "AirQuality"
         capability "FanControl"
         capability "FilterStatus"
+        capability "Switch"
         capability "HealthCheck"
         capability "PowerSource"
         capability "Refresh"
@@ -41,6 +42,7 @@ metadata {
         fingerprint profileId:"0104", endpointId:"01", inClusters:"0000,0003,0004,0005,0202,FC57,FC7C,FC7D", outClusters:"0019,0400,042A", model:"STARKVIND Air purifier table", manufacturer:"IKEA of Sweden"
         
         // Attributes for capability.AirPurifier
+        attribute "airQuality", "enum", ["good", "moderate", "unhealthy for sensitive groups", "unhealthy", "hazardous"]
         attribute "filterUsage", "number"
         attribute "pm25", "number"
         attribute "auto", "enum", ["on", "off"]
@@ -132,8 +134,8 @@ metadata {
 
 // Called when the device is first added
 def installed() {
-    Log.info "Installing device ..."
-    Log.warn "[IMPORTANT] For battery-powered devices, make sure that you keep your device as close as you can to your Hubitat hub (or any other Zigbee router device) for at least 20 seconds. Otherwise it will successfully pair but it won't work properly!"
+    Log.warn "Installing device ..."
+    Log.warn "[IMPORTANT] For battery-powered devices, make sure that you keep your device as close as you can (less than 2inch / 5cm) to your Hubitat hub for at least 30 seconds. Otherwise the device will successfully pair but it won't work properly!"
 }
 
 // Called when the "Save Preferences" button is clicked
@@ -210,7 +212,7 @@ def healthCheck() {
 // capability.Configuration
 // Note: This method is also called when the device is initially installed
 def configure(auto = false) {
-    Log.info "Configuring device${auto ? " (auto)" : ""} ..."
+    Log.warn "Configuring device${auto ? " (auto)" : ""} ..."
     if (!auto && device.currentValue("powerSource", true) == "battery") {
         Log.warn '[IMPORTANT] Click the "Configure" button immediately after pushing any button on the device in order to first wake it up!'
     }
@@ -262,10 +264,11 @@ def configure(auto = false) {
     cmds += zigbee.readAttribute(0x0000, [0x0001, 0x0003, 0x0004, 0x0005, 0x000A, 0x4000]) // ApplicationVersion, HWVersion, ManufacturerName, ModelIdentifier, ProductCode, SWBuildID
     Utils.sendZigbeeCommands cmds
 
-    Log.info "Configuration done! Refreshing device current state in 10 seconds ..."
+    Log.info "Configuration done; refreshing device current state in 10 seconds ..."
     runIn(10, "tryToRefresh")
 }
 private autoConfigure() {
+    Log.warn "Detected that this device is not properly configured for this driver version (lastCx != ${DRIVER_VERSION})"
     configure(true)
 }
 
@@ -345,10 +348,25 @@ def cycleSpeed() {
     Log.debug "Cycling speed to: ${newSpeed}"
     Utils.sendZigbeeCommands(zigbee.writeAttribute(0xFC7D, 0x0006, 0x20, newSpeed, [mfgCode:"0x117C"]))
 }
+private Integer lerp(ylo, yhi, xlo, xhi, cur) {
+  return Math.round(((cur - xlo) / (xhi - xlo)) * (yhi - ylo) + ylo);
+}
+
+// See: https://en.wikipedia.org/wiki/Air_quality_index#United_States
+private pm25Aqi(pm25) {
+    if (pm25 <=  12.1) return [lerp(  0,  50,   0.0,  12.0, pm25), "good", "green"]
+    if (pm25 <=  35.5) return [lerp( 51, 100,  12.1,  35.4, pm25), "moderate", "gold"]
+    if (pm25 <=  55.5) return [lerp(101, 150,  35.5,  55.4, pm25), "unhealthy for sensitive groups", "darkorange"]
+    if (pm25 <= 150.5) return [lerp(151, 200,  55.5, 150.4, pm25), "unhealthy", "red"]
+    if (pm25 <= 250.5) return [lerp(201, 300, 150.5, 250.4, pm25), "very unhealthy", "purple"]
+    if (pm25 <= 350.5) return [lerp(301, 400, 250.5, 350.4, pm25), "hazardous", "maroon"]
+    if (pm25 <= 500.5) return [lerp(401, 500, 350.5, 500.4, pm25), "hazardous", "maroon"]
+    return [500, "hazardous", "maroon"]
+}
 
 // Implementation for capability.HealthCheck
 def ping() {
-    Log.info "ping ..."
+    Log.warn "ping ..."
     Utils.sendZigbeeCommands(zigbee.readAttribute(0x0000, 0x0000))
     Log.debug "Ping command sent to the device; we'll wait 5 seconds for a reply ..."
     runIn 5, "pingExecute"
@@ -375,7 +393,7 @@ def pingExecute() {
 // Implementation for capability.Refresh
 def refresh(buttonPress = true) {
     if (buttonPress) {
-        Log.info "Refreshing current state ..."
+        Log.warn "Refreshing device current state ..."
         if (device.currentValue("powerSource", true) == "battery") {
             Log.warn '[IMPORTANT] Click the "Refresh" button immediately after pushing any button on the device in order to first wake it up!'
         }
@@ -463,14 +481,18 @@ def parse(String description) {
             // Tried to read the PM 2.5 value when the device is Off
             if (pm25 == 0xFFFF) return
         
-            Utils.sendEvent name:"pm25", value:pm25, unit:"μg/m3", descriptionText:"Particulate matter 2.5 is ${pm25} μg/m3", type:type
+            Utils.sendEvent name:"pm25", value:pm25, unit:"μg/m3", descriptionText:"Fine particulate matter (PM2.5) concentration is ${pm25} μg/m3", type:type
+            def aqi = pm25Aqi(pm25)
+            Utils.sendEvent name:"airQualityIndex", value:aqi[0], descriptionText:"Calculated Air Quality Index = ${aqi[0]}", type:type
+            Utils.sendEvent name:"airQuality", value:"<span style=\"color:${aqi[2]}\">${aqi[1]}</span>", descriptionText:"Calculated Air Quality = ${aqi[1]}", type:type
+        
             return Utils.processedZclMessage("${msg.commandInt == 0x0A ? "Report" : "Read"} Attributes Response", "PM25Measurement=${pm25} μg/m3")
         
         // Report/Read Attributes: FilterRunTime
         case { contains it, [clusterInt:0xFC7D, commandInt:0x0A, attrInt:0x0000] }:
         case { contains it, [clusterInt:0xFC7D, commandInt:0x01, attrInt:0x0000] }:
             Integer runTime = Integer.parseInt(msg.value, 16)
-            Float filterUsage = Math.round(runTime * 10000 / (Integer.parseInt(filterLifeTime) * 1440)) / 100.0;
+            Integer filterUsage = Math.floor(runTime * 100 / (Integer.parseInt(filterLifeTime) * 1440));
             Utils.sendEvent name:"filterUsage", value:filterUsage, unit:"%", descriptionText:"Filter usage is ${filterUsage}%", type:type
             return Utils.processedZclMessage("${msg.commandInt == 0x0A ? "Report" : "Read"} Attributes Response", "FilterRunTime=${runTime} min, FilterLifeTime=${filterLifeTime} days")
         
@@ -557,7 +579,7 @@ def parse(String description) {
         
         // Events for capability.HealthCheck
         case { contains it, [clusterInt:0x0000, attrInt:0x0000] }:
-            return Log.info("... pong")
+            return Log.warn("... pong")
         
         // Read Attributes Reponse: PowerSource
         case { contains it, [clusterInt:0x0000, commandInt:0x01, attrInt:0x0007] }:
@@ -613,7 +635,7 @@ def parse(String description) {
 
         // Device_annce: Welcome back! let's sync state.
         case { contains it, [endpointInt:0x00, clusterInt:0x0013, commandInt:0x00] }:
-            Log.info "Rejoined the Zigbee mesh! Refreshing device current state in 3 seconds ..."
+            Log.warn "Rejoined the Zigbee mesh; refreshing device state in 3 seconds ..."
             return runIn(3, "tryToRefresh")
 
         // Read Attributes Response (Basic cluster)
@@ -625,7 +647,7 @@ def parse(String description) {
 
         // Mgmt_Leave_rsp
         case { contains it, [endpointInt:0x00, clusterInt:0x8034, commandInt:0x00] }:
-            return Log.info("Device is leaving the Zigbee mesh. See you later, Aligator!")
+            return Log.warn("Device is leaving the Zigbee mesh. See you later, Aligator!")
 
         // Ignore the following Zigbee messages
         case { contains it, [commandInt:0x0A] }:                                       // ZCL: Attribute report we don't care about (configured by other driver)
@@ -641,7 +663,7 @@ def parse(String description) {
         // Unexpected Zigbee message
         // ---------------------------------------------------------------------------------------------------------------
         default:
-            Log.warn "Sent unexpected Zigbee message: description=${description}, msg=${msg}"
+            Log.error "Sent unexpected Zigbee message: description=${description}, msg=${msg}"
     }
 }