+++ ikea-zigbee-drivers/src/capabilities/AirPurifier.groovy
@@ -1,8 +1,9 @@
 {{!--------------------------------------------------------------------------}}
 {{# @definition }}
-capability "Switch"
+capability "AirQuality"
 capability "FanControl"
 capability "FilterStatus"
+capability "Switch"
 {{/ @definition }}
 {{!--------------------------------------------------------------------------}}
 {{# @fields }}
@@ -16,6 +17,7 @@ capability "FilterStatus"
 {{# @attributes }}
 
 // Attributes for capability.AirPurifier
+attribute "airQuality", "enum", ["good", "moderate", "unhealthy for sensitive groups", "unhealthy", "hazardous"]
 attribute "filterUsage", "number"
 attribute "pm25", "number"
 attribute "auto", "enum", ["on", "off"]
@@ -154,6 +156,21 @@ def cycleSpeed() {
     Log.debug "Cycling speed to: ${newSpeed}"
     Utils.sendZigbeeCommands(zigbee.writeAttribute(0xFC7D, 0x0006, 0x20, newSpeed, [mfgCode:"0x117C"]))
 }
+private Integer lerp(ylo, yhi, xlo, xhi, cur) {
+  return Math.round(((cur - xlo) / (xhi - xlo)) * (yhi - ylo) + ylo);
+}
+
+// See: https://en.wikipedia.org/wiki/Air_quality_index#United_States
+private pm25Aqi(pm25) {
+    if (pm25 <=  12.1) return [lerp(  0,  50,   0.0,  12.0, pm25), "good", "green"]
+    if (pm25 <=  35.5) return [lerp( 51, 100,  12.1,  35.4, pm25), "moderate", "gold"]
+    if (pm25 <=  55.5) return [lerp(101, 150,  35.5,  55.4, pm25), "unhealthy for sensitive groups", "darkorange"]
+    if (pm25 <= 150.5) return [lerp(151, 200,  55.5, 150.4, pm25), "unhealthy", "red"]
+    if (pm25 <= 250.5) return [lerp(201, 300, 150.5, 250.4, pm25), "very unhealthy", "purple"]
+    if (pm25 <= 350.5) return [lerp(301, 400, 250.5, 350.4, pm25), "hazardous", "maroon"]
+    if (pm25 <= 500.5) return [lerp(401, 500, 350.5, 500.4, pm25), "hazardous", "maroon"]
+    return [500, "hazardous", "maroon"]
+}
 {{/ @implementation }}
 {{!--------------------------------------------------------------------------}}
 {{# @updated }}
@@ -219,14 +236,18 @@ case { contains it, [clusterInt:0xFC7D, commandInt:0x01, attrInt:0x0004] }:
     // Tried to read the PM 2.5 value when the device is Off
     if (pm25 == 0xFFFF) return
 
-    Utils.sendEvent name:"pm25", value:pm25, unit:"μg/m3", descriptionText:"Particulate matter 2.5 is ${pm25} μg/m3", type:type
+    Utils.sendEvent name:"pm25", value:pm25, unit:"μg/m3", descriptionText:"Fine particulate matter (PM2.5) concentration is ${pm25} μg/m3", type:type
+    def aqi = pm25Aqi(pm25)
+    Utils.sendEvent name:"airQualityIndex", value:aqi[0], descriptionText:"Calculated Air Quality Index = ${aqi[0]}", type:type
+    Utils.sendEvent name:"airQuality", value:"<span style=\"color:${aqi[2]}\">${aqi[1]}</span>", descriptionText:"Calculated Air Quality = ${aqi[1]}", type:type
+
     return Utils.processedZclMessage("${msg.commandInt == 0x0A ? "Report" : "Read"} Attributes Response", "PM25Measurement=${pm25} μg/m3")
 
 // Report/Read Attributes: FilterRunTime
 case { contains it, [clusterInt:0xFC7D, commandInt:0x0A, attrInt:0x0000] }:
 case { contains it, [clusterInt:0xFC7D, commandInt:0x01, attrInt:0x0000] }:
     Integer runTime = Integer.parseInt(msg.value, 16)
-    Float filterUsage = Math.round(runTime * 10000 / (Integer.parseInt(filterLifeTime) * 1440)) / 100.0;
+    Integer filterUsage = Math.floor(runTime * 100 / (Integer.parseInt(filterLifeTime) * 1440));
     Utils.sendEvent name:"filterUsage", value:filterUsage, unit:"%", descriptionText:"Filter usage is ${filterUsage}%", type:type
     return Utils.processedZclMessage("${msg.commandInt == 0x0A ? "Report" : "Read"} Attributes Response", "FilterRunTime=${runTime} min, FilterLifeTime=${filterLifeTime} days")