+++ ikea-zigbee-drivers/src/capabilities/Refresh.groovy
@@ -8,7 +8,7 @@ capability "Refresh"
 // Implementation for capability.Refresh
 def refresh(buttonPress = true) {
     if (buttonPress) {
-        Log.info "Refreshing current state ..."
+        Log.warn "Refreshing device current state ..."
         if (device.currentValue("powerSource", true) == "battery") {
             Log.warn '[IMPORTANT] Click the "Refresh" button immediately after pushing any button on the device in order to first wake it up!'
         }