+++ ikea-zigbee-drivers/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.1] - 2023-12-12
+
+### Fixed
+- E2006: Display filter usage (%) with no decimals
+- All: Adjusted some log levels (info -> warn)
+
+### Added
+- E2006: Add "airQuality" (enum) attribute based on the reported PM 2.5 value (US AQI standard)
+- E2006: Add "airQualityIndex" (0..500) attribute based on the reported PM 2.5 value (US AQI standard)
+
 ## [3.5.0] - 2023-12-11
 
 ### Added